chore: soundfont code cleanup
